Added printPublicTweets("")
You guys can look at the output. You can leave it the way I have it and
delete the commented parts (i think it looks better the way it is now,
but it's up to you guys), or you can use the commented stuff, which will
make it padded, and put the tweet and the username on the same line.