Finished changePassword(user) and.....
... wrote viewTweet(user) and printProfile(user), which we talked about
a little bit in class today. I also added an option in the logged in
menu for the user to view their own profile, and just added the
viewProfile method there. We may want to clarify the wording a little
for our "V: View Tweets" and " P: View Public Tweets" options in the
menu.