Added tweetAt() (respond)/ fixed viewProfile()
Tweet at (which used to be respond to tweet would work if we could
figure out what to put as the 4th parameter so that it'll write
correctly to tweet.dat. We may have to just scrap the whole hashtag
thing (it's an A option anyway) if we can't figure that out. I tried
null as the 4th parameter and it doesn't work. I also changed
viewProfile so that we can use it for other users too and i added it to
your code Michelle where you said to add it.